new header
Implement new header for MDN Web Docs

fix #1369
