--- client/src/search.scss
@@ -1,112 +1,0 @@
-@import "include-media/dist/_include-media.scss";
-
-$search-border-color: #ccc;
-$search-border-color-focus: #0096bfab;
-
-.search-widget {
-  position: relative;
-  z-index: 12;
-
-  input {
-    width: 400px;
-    margin-bottom: 0;
-    border-radius: 4px;
-    background-color: #fafafa;
-    border: 1px solid $search-border-color;
-  }
-  input:focus {
-    // box-shadow: 0 0 0 1px #0096bfab;
-    box-shadow: none;
-    // border-top: 1px solid $search-border-color-focus;
-    // border-right: 1px solid $search-border-color-focus;
-    // border-left: 1px solid $search-border-color-focus;
-    border: 1px solid $search-border-color-focus;
-    border-radius: 4px 4px 0 0;
-  }
-  input.has-search-results:focus {
-    border-bottom: unset;
-  }
-
-  // // The trickery of having 2 input fields on top of each other.
-  // div.input-wrapper {
-  //   // position: relative;
-  //   input {
-  //     background-color: transparent;
-  //     position: absolute;
-  //     height: 40px;
-  //   }
-  //   input.background {
-  //     color: #666;
-  //     // position: relative;
-  //   }
-  // }
-
-  p.server-error {
-    margin: 0;
-    color: red;
-    font-size: 80%;
-  }
-
-  .search-results {
-    position: absolute;
-    background-color: white;
-    opacity: 0.98;
-    border: 2px solid black;
-    border-radius: 0 0 4px 4px;
-    width: 305px;
-    top: 38px;
-    left: 20px;
-    border-top: 0;
-
-    // This is copied verbatim from the `.main-menu li ul` in the main styles.
-    // Consistency is nice :)
-    box-shadow: 0 2px 8px 0 rgba(12, 12, 13, 0.1);
-
-    .indexing-warning {
-      color: orange;
-    }
-
-    div {
-      padding: 5px;
-      border-bottom: 1px solid #efefef;
-    }
-    .result-item:hover,
-    .result-item.highlit {
-      background-color: #efefef;
-    }
-    .result-item:hover {
-      cursor: pointer;
-    }
-    div:last-child {
-      border-bottom: none;
-    }
-    mark {
-      background-color: #f3f7bd;
-    }
-    small {
-      overflow-wrap: break-word;
-    }
-
-    a {
-      text-overflow: ellipsis;
-    }
-
-    div.nothing-found {
-      font-style: italic;
-    }
-    div.searchindex-error {
-      color: red;
-    }
-    .fuzzy-engaged {
-      font-size: 80%;
-      font-style: italic;
-    }
-  }
-
-  @include media(">phone", "<=tablet") {
-    .search-results,
-    input {
-      width: 100%;
-    }
-  }
-}
