+++ client/src/stories/atoms/button.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Button } from "../../ui/atoms/button.tsx";
+import { Button } from "../../ui/atoms/button/index.tsx";
 
 import { ReactComponent as Arrow } from "@mdn/dinocons/arrows/arrow.svg";
 import { ReactComponent as Github } from "@mdn/dinocons/brands/github-mark-small.svg";
