--- client/src/ui/molecules/search-header.scss
@@ -1,123 +1,0 @@
-@import "~@mdn/minimalist/sass/vars/color-palette";
-@import "~@mdn/minimalist/sass/vars/layout";
-@import "~@mdn/minimalist/sass/mixins/utils";
-
-$mobile-search-height: 68px;
-$icon-size: 24px;
-
-.header-search {
-  position: absolute;
-  top: -62px;
-  left: 0;
-  background-color: $neutral-600;
-  width: 100vw;
-  height: $mobile-search-height;
-  transition: transform 0.1s ease-in;
-
-  @media #{$mq-tablet-and-up} {
-    position: inherit;
-    top: inherit;
-    right: inherit;
-    justify-self: end;
-    width: inherit;
-    height: 40px;
-    min-width: 80px;
-    max-width: 400px;
-  }
-
-  @media #{$mq-large-desktop-and-up} {
-    max-width: 350px;
-  }
-
-  form {
-    position: relative;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: 0 20px;
-    width: 100%;
-    height: $mobile-search-height;
-
-    @media #{$mq-tablet-and-up} {
-      @include reset-visually-hidden();
-      padding: 2px 8px;
-      border: 2px solid $text-color;
-      border-radius: 20px;
-      height: 40px;
-      min-width: 80px;
-      transition: none;
-      transform: translateY(0);
-    }
-
-    .search-icon {
-      display: none;
-
-      @media #{$mq-tablet-and-up} {
-        display: inline-block;
-        flex: 0 0 25px;
-        fill: $text-color;
-      }
-    }
-  }
-
-  input.search-input-field {
-    flex: 1 1 100px;
-    color: $text-color;
-    padding-right: $icon-size + 16px;
-    border-width: 0;
-    min-width: 60px;
-
-    @media #{$mq-tablet-and-up} {
-      padding-right: inherit;
-    }
-
-    &:focus,
-    &:invalid {
-      border: 0;
-      outline: none;
-      box-shadow: none;
-    }
-  }
-
-  .open-search,
-  .close-search {
-    position: absolute;
-    top: $mobile-search-height + 20px;
-    right: 20px;
-    background-position: 0 0;
-    background-repeat: no-repeat;
-    background-size: $icon-size;
-    width: $icon-size;
-    height: $icon-size;
-
-    @media #{$mq-tablet-and-up} {
-      display: none;
-    }
-  }
-
-  .open-search {
-    background-image: url("~@mdn/dinocons/general/search.svg");
-  }
-
-  .close-search {
-    top: $icon-size;
-    background-image: url("~@mdn/dinocons/general/close.svg");
-  }
-
-  &.show-form {
-    position: absolute;
-    transform: translateY($mobile-search-height);
-    z-index: 1;
-
-    @media #{$mq-tablet-and-up} {
-      position: inherit;
-      width: auto;
-    }
-
-    form {
-      @media #{$mq-tablet-and-up} {
-        transform: translateY(0);
-      }
-    }
-  }
-}
