+++ client/src/ui/molecules/search/index.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+
+import { SearchNavigateWidget } from "../../../search";
+
+import "./index.scss";
+
+export default function Search() {
+  return (
+    <div className="header-search">
+      <SearchNavigateWidget />
+    </div>
+  );
+}
