--- client/src/ui/organisms/header.scss
@@ -1,78 +1,0 @@
-@import "~@mdn/minimalist/sass/vars/layout";
-
-.page-header {
-  padding: $base-spacing;
-  position: relative;
-
-  @supports (display: grid) {
-    display: grid;
-    gap: $base-spacing;
-    grid-template-columns: 40px 2fr 80px;
-
-    @media #{$mq-tablet-and-up} {
-      align-items: center;
-      grid-template-columns: 220px 2fr 80px;
-    }
-
-    @media #{$mq-large-desktop-and-up} {
-      grid-template-columns: 220px 2fr 350px 80px;
-    }
-
-    .logo,
-    .main-nav,
-    .signin-link,
-    .auth-container {
-      grid-row: 1/2;
-    }
-
-    .logo {
-      grid-column: 2/3;
-    }
-
-    .main-nav {
-      grid-column: 1/2;
-    }
-
-    .signin-link,
-    .auth-container {
-      grid-column: 3/4;
-    }
-
-    @media #{$mq-tablet-and-up} {
-      .logo {
-        grid-column: 1/2;
-      }
-
-      .main-nav {
-        grid-column: 1/4;
-        grid-row: 2/3;
-      }
-
-      .header-search {
-        grid-column: 2/3;
-        grid-row: 1/2;
-      }
-
-      .signin-link,
-      .auth-container {
-        grid-column: 3/4;
-      }
-    }
-
-    @media #{$mq-large-desktop-and-up} {
-      .main-nav {
-        grid-column: 2/3;
-        grid-row: 1/2;
-      }
-
-      .header-search {
-        grid-column: 3/4;
-      }
-
-      .signin-link,
-      .auth-container {
-        grid-column: 4/5;
-      }
-    }
-  }
-}
