--- client/src/ui/organisms/header.tsx
@@ -1,19 +1,0 @@
-import * as React from "react";
-
-import Login from "../../header/login";
-import { Logo } from "../atoms/logo";
-import MainMenu from "../molecules/main-menu";
-import Search from "../molecules/search-header";
-
-import "./header.scss";
-
-export function Header() {
-  return (
-    <header className="page-header">
-      <Logo />
-      <MainMenu />
-      <Search />
-      <Login />
-    </header>
-  );
-}
