+++ client/src/utils/auth-link.ts
